port over statement-table and validation
